+++ lib/build_bootstrap_structure.py
@@ -1,5 +1,6 @@
 from string import Template
 
+from lib.build_bootstrap_release_planning import build_bootstrap_release_planning_tab
 from lib.build_plotly_analyse import process_analyse
 from lib.build_plotly_bandwidth import process_bandwidth
 from lib.file import read_plotly
@@ -129,214 +130,6 @@ def build_bootstrap_progress(a_start, a_course, a_results, a_templates, a_labels
              'students': students_html_string})
         progress_html_string += level_html_string
     return progress_html_string
-
-
-def write_level_serie(a_course, a_templates, level_serie, a_file_name):
-    status_list_html_string = ""
-    for status_id in level_serie.status:
-        level = level_serie.status[status_id]
-        status_list_html_string += a_templates["level_status"].substitute(
-            {'id': status_id,
-             'label': level.label,
-             'color': level.color})
-    value_list_html_string = ""
-    for grade_id in level_serie.grades:
-        grades = level_serie.grades[grade_id]
-        value_list_html_string += a_templates["level_level"].substitute(
-            {'id': grade_id,
-             'label': grades.label,
-             'color': grades.color,
-             'fraction': grades.fraction,
-             'value': grades.value})
-    file_html_string = a_templates["level_serie_index"].substitute(
-
-        {'level_serie_name': level_serie.name,
-         'status_list': status_list_html_string,
-         'value_list': value_list_html_string})
-    with open(a_file_name, mode='w', encoding="utf-8") as file_list:
-        file_list.write(file_html_string)
-
-
-def write_release_planning_index(course, templates, assignment_group, file_name, file_name_bandwidth):
-    bandwith_html_string = '<h2 class="mt-2">Opbouw</h2>' + read_plotly(file_name_bandwidth)
-    release_planning_html_string = write_release_planning(course, templates, assignment_group)
-    release_planning_index_html_string = templates['release_planning_index'].substitute(
-        {
-            'assignment_group_name': assignment_group.name,
-            'assignment_group_id': str(assignment_group.id),
-            'total_points': int(assignment_group.total_points),
-            'lower_points': assignment_group.lower_points,
-            'upper_points': assignment_group.upper_points,
-            'strategie': assignment_group.strategy,
-            'planning': release_planning_html_string,
-            'flow': bandwith_html_string
-        }
-    )
-
-    with open(file_name, mode='w', encoding="utf-8") as file_release_planning_index:
-        file_release_planning_index.write(release_planning_index_html_string)
-    return
-
-
-def write_learning_outcome_index(course, templates, learning_outcome, file_name):
-    write_learning_outcome_assignments_html_string = get_learning_outcome_assignments(course, templates, learning_outcome)
-    write_learning_outcome_index_html_string = templates['learning_outcome_index'].substitute(
-        {
-            'learning_outcome_short': learning_outcome.short,
-            'learning_outcome_description': learning_outcome.description,
-            'assignment_sequences': write_learning_outcome_assignments_html_string
-        }
-    )
-
-    with open(file_name, mode='w', encoding="utf-8") as file_learning_outcome_index:
-        file_learning_outcome_index.write(write_learning_outcome_index_html_string)
-    return
-
-
-def get_learning_outcome_assignments(a_course, a_templates, a_learning_outcome):
-    html_string = ""
-    portfolio_items = []
-    for assignment_sequence_tag in a_learning_outcome.assignment_sequences:
-        assignment_sequence = a_course.find_assignment_sequence(assignment_sequence_tag)
-        portfolio_items.append({
-                'portfolio_item': assignment_sequence.name,
-                'tag': assignment_sequence.tag,
-                'assignment_date': get_date_time_loc(assignment_sequence.get_date()),
-                'assignment_day': assignment_sequence.get_day(),
-                'points': assignment_sequence.points,
-                'count': len(assignment_sequence.assignments)
-            })
-    portfolio_items = sorted(portfolio_items, key=lambda a: a['assignment_day'])
-    for portfolio_item in portfolio_items:
-        html_string += a_templates['learning_outcome_portfolio_item'].substitute(portfolio_item)
-    return html_string
-
-def write_release_planning(a_course, a_templates, a_assignment_group):
-    list_html_string = ""
-    for assignment_sequence in a_assignment_group.assignment_sequences:
-        messages_html_string = ""
-        assignment_sequence_html_string = ""
-        # print(assignment_sequence.name)
-        for assignment in assignment_sequence.assignments:
-            url = "https://canvas.hu.nl/courses/" + str(a_course.canvas_id) + "/assignments/" + str(assignment.id)
-            rubric_points = 0
-            rubric_count = 0
-            rubrics_html_string = "<ul>"
-            for criterion in assignment.rubrics:
-                rubrics_html_string += "<li>" + str(round(criterion.points, 2)) + " ptn, " + criterion.description
-                rubric_points += criterion.points
-                rubric_count += 1
-                rubrics_html_string += "<ul>"
-                for rating in criterion.ratings:
-                    rubrics_html_string += "<li>" + str(
-                        round(rating.points, 2)) + " ptn, " + rating.description + "</li>"
-                rubrics_html_string += "</ul></li>"
-            rubrics_html_string += "<ul>"
-            if rubric_count == 0:
-                rubrics_str = str(round(0, 2))
-                rubrics_html_string = "Geen criteria"
-            else:
-                rubrics_str = str(round(rubric_points, 2))
-
-            assignment_sequence_html_string += a_templates["assignment"].substitute({'assignment_name': assignment.name,
-                                                                                     'assignment_unlock_date': get_date_time_loc(
-                                                                                         assignment.unlock_date),
-                                                                                     'assignment_lock_date': get_date_time_loc(
-                                                                                         assignment.assignment_date),
-                                                                                     'assignment_grading_type': assignment.grading_type,
-                                                                                     'assignment_points': assignment.points,
-                                                                                     'rubrics_points': rubrics_str,
-                                                                                     'rubrics': rubrics_html_string,
-                                                                                     'url': url})
-
-            for message in assignment.messages:
-                messages_html_string += a_templates["message"].substitute({'message': message})
-
-        list_html_string += a_templates["assignment_sequence"].substitute({'assignment_name': assignment_sequence.name,
-                                                                           'assignment_tag': assignment_sequence.tag,
-                                                                           'assignment_grading_type': assignment_sequence.grading_type,
-                                                                           'assignment_points': assignment_sequence.points,
-                                                                           'messages': messages_html_string,
-                                                                           'assignments': assignment_sequence_html_string})
-    release_planning_html_string = a_templates["release_planning_list"].substitute(
-        {'assignment_group': a_assignment_group.name + " " + str(a_assignment_group.id),
-         'total_points': int(a_assignment_group.total_points), 'lower_points': a_assignment_group.lower_points,
-         'upper_points': a_assignment_group.upper_points, 'strategie': a_assignment_group.strategy,
-         'assignments': list_html_string})
-    return release_planning_html_string
-
-
-def get_perspective_assignment_group_release_planning(a_course, a_perspective, a_assignment_group_id, a_templates):
-    assignment_group = a_course.find_assignment_group(a_assignment_group_id)
-    file_name_levels = ".//general//level_serie_" + str(a_perspective.levels) + ".html"
-    file_name_group = ".//general//release_planning_" + str(a_assignment_group_id) + ".html"
-    print("BBS31 -", file_name_group)
-
-    return a_templates["release_planning_perspective"].substitute(
-        {'url_levels': file_name_levels,
-         'url_group': file_name_group,
-         'perspective_name': a_perspective.title,
-         'levels': a_perspective.levels,
-         'assignment_group_name': assignment_group.name,
-         'role': assignment_group.role})
-
-
-def build_bootstrap_release_planning_tab(a_instance, a_course, a_templates, level_serie_collection):
-    html_string = ""
-    for level_serie in level_serie_collection.level_series.values():
-        file_name = a_instance.get_html_path() + "level_serie_" + str(level_serie.name) + ".html"
-        write_level_serie(a_course, a_templates, level_serie, file_name)
-    for assignment_group in a_course.assignment_groups:
-        file_name_bandwidth = a_instance.get_temp_path() + "bandwidth_" + str(assignment_group.id) + ".html"
-        process_bandwidth(a_course, assignment_group, level_serie_collection, file_name_bandwidth)
-    for assignment_group in a_course.assignment_groups:
-        file_name_bandwidth = a_instance.get_temp_path() + "bandwidth_" + str(assignment_group.id) + ".html"
-        file_name_release_planning = a_instance.get_html_path() + "release_planning_" + str(
-            assignment_group.id) + ".html"
-        write_release_planning_index(a_course, a_templates, assignment_group, file_name_release_planning,
-                                     file_name_bandwidth)
-
-    perspectives_html_string = ""
-    for perspective in a_course.perspectives.values():
-        for assignment_group_id in perspective.assignment_groups:
-            perspectives_html_string += get_perspective_assignment_group_release_planning(a_course, perspective,
-                                                                                          assignment_group_id,
-                                                                                          a_templates)
-    if a_course.level_moments is not None:
-        assignment_group_id = a_course.level_moments.assignment_groups[0]
-        perspectives_html_string += get_perspective_assignment_group_release_planning(a_course, a_course.level_moments,
-                                                                                      assignment_group_id,
-                                                                                      a_templates)
-    if a_course.grade_moments is not None:
-        assignment_group_id = a_course.grade_moments.assignment_groups[0]
-        perspectives_html_string += get_perspective_assignment_group_release_planning(a_course, a_course.grade_moments,
-                                                                                      assignment_group_id,
-                                                                                      a_templates)
-    html_string += a_templates["release_planning"].substitute(
-        {'perspectives': perspectives_html_string})
-    return html_string
-
-
-def build_bootstrap_leeruitkomsten_tab(instance, course, templates, a_level_serie_collection):
-    html_string = ""
-    learning_outcomes_html_string = ""
-    for learning_outcome in course.learning_outcomes:
-        # assignment_html_string = ""
-        # assignment_list = []
-        # for assignment_sequence in assignment_group.assignment_sequences:
-        #     for assignment in assignment_sequence.assignments:
-        #         assignment_list.append(assignment)
-        # assignment_list = sorted(assignment_list, key=lambda a: a.assignment_day)
-        # for assignment in assignment_list:
-        file_name = "general/learning_outcome_" + str(learning_outcome.id).lower() + ".html"
-        print("BBL41 -", file_name)
-        learning_outcomes_html_string += templates["learning_outcome"].substitute(
-                {'url': file_name,
-                 'learning_outcome_short': learning_outcome.short})
-        write_learning_outcome_index(course, templates, learning_outcome, instance.get_html_root_path()+file_name)
-
-    html_string += templates["learning_outcome_card"].substitute({'learning_outcomes': learning_outcomes_html_string})
-    return html_string
 
 
 def build_bootstrap_analyse_tab(instance, a_course, learning_analytics, a_templates, a_level_serie_collection, actual_day):
