+++ runner.py
@@ -41,7 +41,8 @@ if __name__ == "__main__":
         main(sys.argv[1], sys.argv[2])
     else:
         # main("TICT-V1SE1-24_SEP2024", "results_create_event")
-        main("sep24_inno", "results_create_event")
+        # main("sep24_inno", "results_create_event")
+        main("inno2_inno", "results_create_event")
     total_seconds = (get_actual_date() - l_actual_date).seconds
     seconds = total_seconds % 60
     minutes = total_seconds // 60
