+++ templates/learning_outcome/template_learning_outcome_index.html
@@ -36,7 +36,9 @@
                             <h5 class="m-0 font-weight-bold text-primary">$learning_outcome_short</h5>
                         </div>
                         <div class="card-body">
-                            $learning_outcome_description
+                            $learning_outcome_description<br>
+                            Op niveau $up_level<br>
+                            Boven niveau $above_level
                         </div>
                     </div>
                 </div>
